requirements.txt
fastapi
requests
python-dotenv
pytest
pandas
PyJWT
pydantic
pydantic[email]
uvicorn
httpx
